+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to
 - ♿️(frontend) improve accessibility of the IntroSlider carousel #1026
 - ♿️(frontend) add skip link component for keyboard navigation #1019
 - ♿️(frontend) announce mic/camera state to SR on shortcut toggle #1052
+- ♿(frontend) Align the Settings dialog heading level #1051
 
 ### Fixed